Summary: find time to schedule nodes
Details:
As the resource pools is expanded and the number of
active schedules increases, it becomes increasingly
difficult to find a suitable time for the number of
nodes needed for the number of days needed.

With find-available.py, we can pass the node count we want
and the number of days we need, and the schedule.py script
is used repeatedly to determine a set of hosts that are
available for consecutive days for the number of days needed.

Various python libraries are used to make this normally
difficult and complex calculation run fast and provide
the scheduling commands to allocate specific nodes for the
requested use case.